add ssh example and fix proposal layer (#60)
* fix mem size_t

* add ssh example

* add readme

* fix proposal

* fix

* blob shape

* fix

* travis

* fix path

* no test on protobuf2.5

* add blob count api

* fix

* update

* update

+++ src/syncedmem.cpp
@@ -167,12 +167,12 @@ inline std::string MemSize(double size) {
   return os.str();
 }
 
-inline bool ShouldBorrowMem(int has, int wants) {
+inline bool ShouldBorrowMem(size_t has, size_t wants) {
   const int ratio = 2;
   return has / 2 <= wants;
 }
 
-MemBlock MemoryPool::RequestCPU(int size) {
+MemBlock MemoryPool::RequestCPU(size_t size) {
   MemBlock block;
   if (size <= kElementSize) {  // small object <= 128 bytes
     block.device = -1;
@@ -231,7 +231,7 @@ void MemoryPool::ReturnCPU(MemBlock block) {
   }
 }
 
-MemBlock MemoryPool::RequestGPU(int size, int device) {
+MemBlock MemoryPool::RequestGPU(size_t size, int device) {
   MemBlock block;
 #ifdef USE_CUDA
   GpuKey key{device, size};
